#12 Move modalTest to examples folder
